--- deployment/ccip/changeset/active_candidate.go
@@ -1,136 +1,0 @@
-package changeset
-
-import (
-	"fmt"
-
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/smartcontractkit/ccip-owner-contracts/pkg/gethwrappers"
-	"github.com/smartcontractkit/ccip-owner-contracts/pkg/proposal/mcms"
-	"github.com/smartcontractkit/ccip-owner-contracts/pkg/proposal/timelock"
-
-	"github.com/smartcontractkit/chainlink/deployment"
-	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/internal"
-	"github.com/smartcontractkit/chainlink/deployment/common/proposalutils"
-	cctypes "github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/types"
-)
-
-// PromoteAllCandidatesChangeset generates a proposal to call promoteCandidate on the CCIPHome through CapReg.
-// This needs to be called after SetCandidateProposal is executed.
-// TODO: make it conform to the ChangeSet interface.
-func PromoteAllCandidatesChangeset(
-	state CCIPOnChainState,
-	homeChainSel, newChainSel uint64,
-	nodes deployment.Nodes,
-) (deployment.ChangesetOutput, error) {
-	promoteCandidateOps, err := PromoteAllCandidatesForChainOps(
-		state.Chains[homeChainSel].CapabilityRegistry,
-		state.Chains[homeChainSel].CCIPHome,
-		newChainSel,
-		nodes.NonBootstraps(),
-	)
-	if err != nil {
-		return deployment.ChangesetOutput{}, err
-	}
-
-	var (
-		timelocksPerChain = map[uint64]common.Address{
-			homeChainSel: state.Chains[homeChainSel].Timelock.Address(),
-		}
-		proposerMCMSes = map[uint64]*gethwrappers.ManyChainMultiSig{
-			homeChainSel: state.Chains[homeChainSel].ProposerMcm,
-		}
-	)
-	prop, err := proposalutils.BuildProposalFromBatches(
-		timelocksPerChain,
-		proposerMCMSes,
-		[]timelock.BatchChainOperation{{
-			ChainIdentifier: mcms.ChainIdentifier(homeChainSel),
-			Batch:           promoteCandidateOps,
-		}},
-		"promoteCandidate for commit and execution",
-		0, // minDelay
-	)
-	if err != nil {
-		return deployment.ChangesetOutput{}, err
-	}
-	return deployment.ChangesetOutput{
-		Proposals: []timelock.MCMSWithTimelockProposal{
-			*prop,
-		},
-	}, nil
-}
-
-// SetCandidateExecPluginProposal calls setCandidate on the CCIPHome for setting up OCR3 exec Plugin config for the new chain.
-// TODO: make it conform to the ChangeSet interface.
-func SetCandidatePluginChangeset(
-	state CCIPOnChainState,
-	e deployment.Environment,
-	nodes deployment.Nodes,
-	ocrSecrets deployment.OCRSecrets,
-	homeChainSel, feedChainSel, newChainSel uint64,
-	tokenConfig TokenConfig,
-	pluginType cctypes.PluginType,
-) (deployment.ChangesetOutput, error) {
-	ccipOCRParams := DefaultOCRParams(
-		feedChainSel,
-		tokenConfig.GetTokenInfo(e.Logger, state.Chains[newChainSel].LinkToken, state.Chains[newChainSel].Weth9),
-		nil,
-	)
-	newDONArgs, err := internal.BuildOCR3ConfigForCCIPHome(
-		ocrSecrets,
-		state.Chains[newChainSel].OffRamp,
-		e.Chains[newChainSel],
-		nodes.NonBootstraps(),
-		state.Chains[homeChainSel].RMNHome.Address(),
-		ccipOCRParams.OCRParameters,
-		ccipOCRParams.CommitOffChainConfig,
-		ccipOCRParams.ExecuteOffChainConfig,
-	)
-	if err != nil {
-		return deployment.ChangesetOutput{}, err
-	}
-
-	execConfig, ok := newDONArgs[pluginType]
-	if !ok {
-		return deployment.ChangesetOutput{}, fmt.Errorf("missing exec plugin in ocr3Configs")
-	}
-
-	setCandidateMCMSOps, err := SetCandidateOnExistingDon(
-		execConfig,
-		state.Chains[homeChainSel].CapabilityRegistry,
-		state.Chains[homeChainSel].CCIPHome,
-		newChainSel,
-		nodes.NonBootstraps(),
-	)
-	if err != nil {
-		return deployment.ChangesetOutput{}, err
-	}
-
-	var (
-		timelocksPerChain = map[uint64]common.Address{
-			homeChainSel: state.Chains[homeChainSel].Timelock.Address(),
-		}
-		proposerMCMSes = map[uint64]*gethwrappers.ManyChainMultiSig{
-			homeChainSel: state.Chains[homeChainSel].ProposerMcm,
-		}
-	)
-	prop, err := proposalutils.BuildProposalFromBatches(
-		timelocksPerChain,
-		proposerMCMSes,
-		[]timelock.BatchChainOperation{{
-			ChainIdentifier: mcms.ChainIdentifier(homeChainSel),
-			Batch:           setCandidateMCMSOps,
-		}},
-		"SetCandidate for execution",
-		0, // minDelay
-	)
-	if err != nil {
-		return deployment.ChangesetOutput{}, err
-	}
-	return deployment.ChangesetOutput{
-		Proposals: []timelock.MCMSWithTimelockProposal{
-			*prop,
-		},
-	}, nil
-
-}
