+++ integration-tests/smoke/ccip/ccip_fee_boosting_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/smartcontractkit/chainlink-common/pkg/config"
+	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -90,8 +91,16 @@ func Test_CCIPFeeBoosting(t *testing.T) {
 		DestSelector:          destChain,
 		InitialPricesBySource: initialPrices,
 		FeeQuoterDestChain:    changeset.DefaultFeeQuoterDestChainConfig(),
+		TestRouter:            false,
 	}
-	require.NoError(t, changeset.AddLane(e.Env, state, laneCfg, false))
+
+	e.Env, err = commonchangeset.ApplyChangesets(t, e.Env, nil, []commonchangeset.ChangesetApplication{
+		{
+			Changeset: commonchangeset.WrapChangeSet(changeset.AddLanes),
+			Config:    changeset.AddLanesConfig{LaneConfigs: []changeset.LaneConfig{laneCfg}},
+		},
+	})
+	require.NoError(t, err)
 
 	// Update token prices in destination chain FeeQuoter
 	err = updateTokensPrices(e, state, destChain, map[common.Address]*big.Int{
